Add files via upload
Chelsea, here are the initialization files.  Right now in the code it looks like I have turned off the rsinput.dat file in lieu of screen prompts.  However when you go to run these in batch it's better practice to have an initialization file for a model run (i.e. your experiment) so that you can track down any issues that might have stemmed from fat fingering  model parameterization value (at least you have a way of knowing what initial values were).  It can run without the rsinput but it cannot run without the BIOEN files which set the bioenergetics values for each species that we modeled in that project.